% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurora_jl.R
\name{aurora_knn}
\alias{aurora_knn}
\title{Aurora with K-Nearest Neighbors}
\usage{
aurora_knn(Zs, loocv = TRUE, kKNN = 1000, predictions_only = TRUE)
}
\arguments{
\item{Zs}{A N*K matrix of noisy measurements. Each row corresponds to a different units.
The K measurements in the i-th row are the replicate measurements for the i-th unit.}

\item{loocv}{Boolean, whether to choose K by leave-one-out-cross-validation. Defaults to TRUE.}

\item{kKNN}{Integer. If loocv = FALSE, this specifies the number of neighbors to use.
If loocv = TRUE, then K is chosen by LOOCV from the set {1, ..., kKNN}. Defaults to 1000.}

\item{predictions_only}{Boolean, whether to return only vector of predictions, or
to also return the objects containing more informations about the fits. Defaults to TRUE.}
}
\value{
A vector of length N of estimates of the mean of each unit.
}
\description{
Aurora with K-Nearest Neighbors
}
