% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurora_jl.R
\name{aurora_setup}
\alias{aurora_setup}
\title{Setup AuroraR}
\usage{
aurora_setup(installJulia = TRUE, pkg_check = TRUE, ...)
}
\arguments{
\item{installJulia}{logical, whether to install Julia automatically when Julia is not found,
whose default value is TRUE.}

\item{pkg_check}{logical, check for Aurora.jl package and install if necessary.}

\item{...}{arguments passed to \code{JuliaCall::julia_setup}.}
}
\description{
This function initializes Julia and the Aurora.jl package.
The first time will be long since it includes precompilation.
Additionally, this will install Julia and the required packages
if they are missing.
}
\references{
Similar functions are available from other R packages that call JuliaCall,
for example: \cr
https://github.com/SciML/diffeqr/blob/master/R/diffeqr.R \cr
https://github.com/Non-Contradiction/ipoptjlr/blob/master/R/IPOPT.R
}
