% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurora_jl.R
\name{auroralmethods}
\alias{auroralmethods}
\alias{auroral}
\alias{ebcc}
\title{Auroral estimators}
\usage{
auroral(Zs, predictions_only = TRUE)

ebcc(Zs, predictions_only = TRUE)
}
\arguments{
\item{Zs}{A N*K matrix of noisy measurements. Each row corresponds to a different units.
The K measurements in the i-th row are the replicate measurements for the i-th unit.}

\item{predictions_only}{Boolean, whether to return only vector of predictions, or
to also return the objects containing more informations about the fits. Defaults to TRUE.}
}
\value{
A vector of length N of estimates of the mean of each unit.
}
\description{
Auroral estimators
}
\section{Functions}{
\itemize{
\item \code{auroral}: Auroral: Aurora with linear regression

\item \code{ebcc}: Coey-Cunningham: Similar to Aurora with linear regression,
but K-1 replicates are first summarized to their sample mean
}}

