% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{gamma_pareto_sim}
\alias{gamma_pareto_sim}
\title{Simulation: Pareto likelihood, Gamma Prior}
\usage{
gamma_pareto_sim(m, B, pareto_alpha = 3, gamma_shape = 1, gamma_scale = 1)
}
\arguments{
\item{m}{Integer, number of units.}

\item{B}{Integer, number of replicates.}

\item{pareto_alpha}{Numeric, tail index of Pareto distribution (defaults to 3, note
that it needs to be >2 for the variance to exist).}

\item{gamma_shape}{Numeric, shape of the prior Gamma distribution (defaults to 1).}

\item{gamma_scale}{Numeric, scale of the prior Gamma distribution (defaults to 1).}
}
\value{
A list with the following entries:
     \item{m}{The number of units in the simulation.}
     \item{B}{The number of replicates per unit in the simulation.}
     \item{Zs}{A m*B matrix of noisy measurements. Each row corresponds to a different units.
    The K measurements in the i-th row are the replicate measurements for the i-th unit.}
     \item{sigmas}{The noise standard deviation of the measurements of each unit. Can be either a single number (if
        the simulation is homoskedastic) or a vector of length m.}
     \item{...}{Other entries that contains simulation specific parameters.}
}
\description{
mu_i ~ Gamma(gamma_shape, gamma_scale), Z_ij ~ Pareto( mean = mu_i, alpha)
}
