% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_estimators.R
\name{gmleb}
\alias{gmleb}
\title{GMLEB assuming Gaussian average measurements (CLT)}
\usage{
gmleb(Zs, sigma = 1)
}
\arguments{
\item{Zs}{A N*K matrix of noisy measurements. Each row corresponds to a different units.
The K measurements in the i-th row are the replicate measurements for the i-th unit.}

\item{sigma}{Numeric, noise standard deviation (of replicate average, i.e.,
each noisy measurements has a standard deviation equal to sqrt(K)*sigma).}
}
\value{
A vector of length N of estimates of the mean of each unit.
}
\description{
GMLEB assuming Gaussian average measurements (CLT)
}
\references{
This is just a wrapper around the implementation of GLmix in the REBayes package
     (cite REBayes if you use this function).
}
