% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_estimators.R
\name{grouplinear}
\alias{grouplinear}
\title{Group-linear estimator}
\usage{
grouplinear(Zs, sigma = 1, nbreak = floor(nrow(Zs)^(1/3)))
}
\arguments{
\item{Zs}{A N*K matrix of noisy measurements. Each row corresponds to a different units.
The K measurements in the i-th row are the replicate measurements for the i-th unit.}

\item{sigma}{Numeric, noise standard deviation (of replicate average, i.e.,
each noisy measurements has a standard deviation equal to sqrt(K)*sigma).}

\item{nbreak}{Integer, number of breaks for group-linear estimators. Defaults to floor(N^(1/3)).}
}
\value{
A vector of length N of estimates of the mean of each unit.
}
\description{
This code is directly adapted from https://github.com/MaZhuang/grouplinear,
with only minor changes.
}
\references{
"Group-linear empirical Bayes estimates for a heteroscedastic normal mean"
       by A Weinstein, Z Ma, LD Brown, CH Zhang (JASA, 2018)
}
