% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{heteroskedasticloc}
\alias{heteroskedasticloc}
\alias{wmbz_example_a_sim}
\alias{wmbz_example_c_sim}
\alias{wmbz_example_f_sim}
\title{Heteroskedastic location family simulations}
\usage{
wmbz_example_a_sim(m, B, sigma_squared_lower, sigma_squared_upper, A = 1)

wmbz_example_c_sim(m, B, sigma_squared_lower, sigma_squared_upper)

wmbz_example_f_sim(m, B, sigma_squared_lower, sigma_squared_upper)
}
\arguments{
\item{m}{Integer, number of units.}

\item{B}{Integer, number of replicates.}

\item{sigma_squared_lower}{Numeric, lower limit of noise variance.}

\item{sigma_squared_upper}{Numeric, upper limit of noise variance.}

\item{A}{Numeric, Prior noise variance.}
}
\value{
A list with the following entries:
     \item{m}{The number of units in the simulation.}
     \item{B}{The number of replicates per unit in the simulation.}
     \item{Zs}{A m*B matrix of noisy measurements. Each row corresponds to a different units.
    The K measurements in the i-th row are the replicate measurements for the i-th unit.}
     \item{sigmas}{The noise standard deviation of the measurements of each unit. Can be either a single number (if
        the simulation is homoskedastic) or a vector of length m.}
     \item{...}{Other entries that contains simulation specific parameters.}
}
\description{
The variance of the replicates of unit i vary across units (randomly), as follows:
}
\details{
sigma_i^2 ~ U[sigma_squared_lower, sigma_squared_upper],

where B is the number of replicates per unit and
sigma_squared_lower, sigma_squared_upper are simulation parameters.

Note that sigma_i^2 is the variance of the replicate average, i.e.,
i.e.,  each noisy measurements has a standard deviation equal to sqrt(B)*sigma_i)
}
\section{Functions}{
\itemize{
\item \code{wmbz_example_a_sim}: Here mu_i ~ N(0, A). Then Z_ij ~ N(mu_i, B*sigma_i^2).

\item \code{wmbz_example_c_sim}: Here mu_i = sigma_i^2 deterministically. Then Z_ij ~ N(mu_i, B*sigma_i^2).

\item \code{wmbz_example_f_sim}: Here mu_i = sigma_i^2 deterministically. Then Z_ij = mu_i + sqrt(3B)*sigma_i*U[-1, 1].
}}

\references{
The simulations here in are adapted from the paper
      "Group-linear empirical Bayes estimates for a heteroscedastic normal mean"
       by A Weinstein, Z Ma, LD Brown, CH Zhang (JASA, 2018)
}
