% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_estimators.R
\name{james_stein}
\alias{james_stein}
\title{James-Stein estimator}
\usage{
james_stein(Zs, sigma = 1, positive_part = TRUE)
}
\arguments{
\item{Zs}{A N*K matrix of noisy measurements. Each row corresponds to a different units.
The K measurements in the i-th row are the replicate measurements for the i-th unit.}

\item{sigma}{Numeric, noise standard deviation (of replicate average, i.e.,
each noisy measurements has a standard deviation equal to sqrt(K)*sigma).}

\item{positive_part}{Bool, whether to use the positive part James-Stein estimator
which uniformly dominates the James-Stein estimator (defaults to TRUE).}
}
\value{
A vector of length N of estimates of the mean of each unit.
}
\description{
James-Stein estimator
}
