% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_estimators.R
\name{simpleloc}
\alias{simpleloc}
\alias{location_mean}
\alias{location_median}
\alias{location_midrange}
\alias{pareto_mle}
\title{Simple location estimators}
\usage{
location_mean(Zs)

location_median(Zs)

location_midrange(Zs)

pareto_mle(Zs)
}
\arguments{
\item{Zs}{A N*K matrix of noisy measurements. Each row corresponds to a different units.
The K measurements in the i-th row are the replicate measurements for the i-th unit.}
}
\value{
A vector of length N of estimates of the mean of each unit.
}
\description{
Simple location estimators
}
\section{Functions}{
\itemize{
\item \code{location_mean}: row-wise sample mean (average)

\item \code{location_median}: row-wise median

\item \code{location_midrange}: row-wise midrange

\item \code{pareto_mle}: Maximum likelihood estimator of mean,
assuming each row consists of i.i.d. samples from the Pareto distribution
with unknown parameters.
}}

