% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_estimators.R
\name{sure_grandmean}
\alias{sure_grandmean}
\title{SURE shrinkage in heteroskedastic Gaussian problems towards grand mean}
\usage{
sure_grandmean(Zs, sigma = 1)
}
\arguments{
\item{Zs}{A N*K matrix of noisy measurements. Each row corresponds to a different units.
The K measurements in the i-th row are the replicate measurements for the i-th unit.}

\item{sigma}{Numeric, noise standard deviation (of replicate average, i.e.,
each noisy measurements has a standard deviation equal to sqrt(K)*sigma).}
}
\value{
A vector of length N of estimates of the mean of each unit.
}
\description{
This code is directly adapted from https://github.com/MaZhuang/grouplinear,
with minor changes.
}
\references{
Xie, X., Kou, S. C., & Brown, L. D. (2012). SURE estimates for a heteroscedastic hierarchical model.
         Journal of the American Statistical Association, 107(500), 1465-1479.
}
