% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{threecomponentloc}
\alias{threecomponentloc}
\alias{three_component_normal_sim}
\alias{three_component_laplace_sim}
\alias{three_component_uniform_sim}
\title{Homoskedastic location family simulations with Three component prior}
\usage{
three_component_normal_sim(m, B, sigma, A = 9)

three_component_laplace_sim(m, B, sigma, A = 9)

three_component_uniform_sim(m, B, sigma, A = 9)
}
\arguments{
\item{m}{Integer, number of units.}

\item{B}{Integer, number of replicates.}

\item{sigma}{Numeric, noise standard deviation (of replicate average, i.e.,
each noisy measurements has a standard deviation equal to sqrt(B)*sigma).}

\item{A}{Numeric, prior variance (defaults to 9). Note that the c on which
G puts its mass is equal to sqrt(1.5*A).}
}
\value{
A list with the following entries:
     \item{m}{The number of units in the simulation.}
     \item{B}{The number of replicates per unit in the simulation.}
     \item{Zs}{A m*B matrix of noisy measurements. Each row corresponds to a different units.
    The K measurements in the i-th row are the replicate measurements for the i-th unit.}
     \item{sigmas}{The noise standard deviation of the measurements of each unit. Can be either a single number (if
        the simulation is homoskedastic) or a vector of length m.}
     \item{...}{Other entries that contains simulation specific parameters.}
}
\description{
Here the prior assigns equal probability to -c, 0, +c, where c is specified
through the parameter A.
}
\section{Functions}{
\itemize{
\item \code{three_component_normal_sim}: Three component prior, Normal likelihood

\item \code{three_component_laplace_sim}: Three component prior, Laplace likelihood

\item \code{three_component_uniform_sim}: Three component prior, Rectangular (Uniform) likelihood
}}

