% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{uniform_pareto_sim}
\alias{uniform_pareto_sim}
\title{Simulation: Pareto likelihood, Uniform Prior}
\usage{
uniform_pareto_sim(m, B, pareto_alpha = 3, unif_lb = 0.1, unif_ub = 1)
}
\arguments{
\item{m}{Integer, number of units.}

\item{B}{Integer, number of replicates.}

\item{pareto_alpha}{Numeric, tail index of Pareto distribution (defaults to 3, note
that it needs to be >2 for the variance to exist).}

\item{unif_lb}{Numeric, lower support point of the prior Uniform distribution (defaults to 0.1).}

\item{unif_ub}{Numeric, upper support point of the prior Uniform distribution (defaults to 1).}
}
\value{
A list with the following entries:
     \item{m}{The number of units in the simulation.}
     \item{B}{The number of replicates per unit in the simulation.}
     \item{Zs}{A m*B matrix of noisy measurements. Each row corresponds to a different units.
    The K measurements in the i-th row are the replicate measurements for the i-th unit.}
     \item{sigmas}{The noise standard deviation of the measurements of each unit. Can be either a single number (if
        the simulation is homoskedastic) or a vector of length m.}
     \item{...}{Other entries that contains simulation specific parameters.}
}
\description{
mu_i ~ Uniform[unif_lb, unif_ub], Z_ij ~ Pareto( mean = mu_i, alpha)
}
